refactor: restructure internals
