--- src/index.test.ts
@@ -1,7 +1,0 @@
-import { describe, expect, it } from "vitest";
-
-describe("index", () => {
-  it("should work", () => {
-    expect(true).toBe(true);
-  });
-});
